requirements.txt
streamlit>=1.32.0
pyyaml>=6.0
google-cloud-bigquery>=3.0.0
